requirements.txt
# Flask Framework
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# Database
pymongo==4.6.1
dnspython==2.4.2

# Authentication
PyJWT==2.8.0

# Machine Learning
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2

# Environment Variables
python-dotenv==1.0.0

# CORS (if needed for API)
flask-cors==4.0.0
